refactor: split attention forward pipeline
