+++ crates/bitnet-transformer/src/lib.rs
@@ -11,6 +11,8 @@ use candle_core::{DType, Device, Module, Tensor};
 use candle_nn::{LayerNorm, Linear, VarBuilder};
 use std::sync::OnceLock;
 
+mod attention_forward;
+
 #[derive(Clone, Debug)]
 struct QwenTraceConfig {
     path: Option<std::path::PathBuf>,
@@ -580,338 +582,9 @@ impl MultiHeadAttention {
         })
     }
 
-    pub fn forward(
-        &self,
-        x: &Tensor,
-        kv_cache: Option<&mut LayerKVCache>,
-        raw_tensors: &std::collections::HashMap<String, Tensor>,
-    ) -> Result<Tensor> {
-        let (batch_size, seq_len, _) = x.dims3()?;
-
-        // PATCH 3: Project to Q, K, V separately (NOT fused QKV)
-        // This is the correct implementation - separate projections ensure proper shape handling
-        // Q: [B, T, hidden] -> [B, T, n_heads * head_dim] -> [B, n_heads, T, head_dim]
-        // K: [B, T, hidden] -> [B, T, n_kv_heads * head_dim] -> [B, n_kv_heads, T, head_dim]
-        // V: [B, T, hidden] -> [B, T, n_kv_heads * head_dim] -> [B, n_kv_heads, T, head_dim]
-        let q_proj_out = self.apply_linear(x, &self.q_proj, "q_proj", raw_tensors)?;
-        let k_proj_out = self.apply_linear(x, &self.k_proj, "k_proj", raw_tensors)?;
-        let v_proj_out = self.apply_linear(x, &self.v_proj, "v_proj", raw_tensors)?;
-        if qwen_trace_layer_enabled(self.layer_idx) {
-            qwen_trace_tensor("attention.q_proj", Some(self.layer_idx), &q_proj_out)?;
-            qwen_trace_tensor("attention.k_proj", Some(self.layer_idx), &k_proj_out)?;
-            qwen_trace_tensor("attention.v_proj", Some(self.layer_idx), &v_proj_out)?;
-        }
-
-        // Probe A3: Q/K/V projection RMS (layer 0, step 0 only)
-        if trace_rms_enabled() && self.layer_idx == 0 {
-            static PROJ_LOGGED: std::sync::Once = std::sync::Once::new();
-            PROJ_LOGGED.call_once(|| {
-                let _ = (|| -> candle_core::Result<()> {
-                    let q_vec = q_proj_out.flatten_all()?.to_vec1::<f32>()?;
-                    let q_rms = (q_vec.iter().map(|x| x * x).sum::<f32>()
-                        / q_vec.len().max(1) as f32)
-                        .sqrt();
-                    let k_vec = k_proj_out.flatten_all()?.to_vec1::<f32>()?;
-                    let k_rms = (k_vec.iter().map(|x| x * x).sum::<f32>()
-                        / k_vec.len().max(1) as f32)
-                        .sqrt();
-                    let v_vec = v_proj_out.flatten_all()?.to_vec1::<f32>()?;
-                    let v_rms = (v_vec.iter().map(|x| x * x).sum::<f32>()
-                        / v_vec.len().max(1) as f32)
-                        .sqrt();
-                    eprintln!(
-                        "trace: q_proj_rms={:.6} k_proj_rms={:.6} v_proj_rms={:.6}",
-                        q_rms, k_rms, v_rms
-                    );
-                    Ok(())
-                })();
-            });
-        }
-
-        // Tracepoint 3: Q projection output (layer-specific)
-        #[cfg(feature = "trace")]
-        {
-            let trace_name = format!("t0/blk{}/q_proj", self.layer_idx);
-            bitnet_trace::dump_trace(
-                &trace_name,
-                &q_proj_out,
-                Some(0),
-                Some(self.layer_idx as isize),
-                Some("q_proj"),
-            )
-            .map_err(BitNetError::from)?;
-        }
-
-        let q = q_proj_out
-            .reshape(&[batch_size, seq_len, self.n_heads, self.head_dim])?
-            .transpose(1, 2)?; // [B, Hq, T, D]
-
-        let k = k_proj_out
-            .reshape(&[batch_size, seq_len, self.n_kv_heads, self.head_dim])?
-            .transpose(1, 2)?; // [B, HKV, T, D]
-
-        let v = v_proj_out
-            .reshape(&[batch_size, seq_len, self.n_kv_heads, self.head_dim])?
-            .transpose(1, 2)?; // [B, HKV, T, D]
-
-        // Debug Q, K, V projections
-        dbg_stats("Q", &q)?;
-        dbg_stats("K", &k)?;
-        dbg_stats("V", &v)?;
-
-        let q = if let Some(norm) = &self.q_norm {
-            let normalized = norm.forward(&q)?;
-            if qwen_trace_layer_enabled(self.layer_idx) {
-                qwen_trace_tensor("attention.q_norm", Some(self.layer_idx), &normalized)?;
-            }
-            normalized
-        } else {
-            q
-        };
-
-        let k = if let Some(norm) = &self.k_norm {
-            let normalized = norm.forward(&k)?;
-            if qwen_trace_layer_enabled(self.layer_idx) {
-                qwen_trace_tensor("attention.k_norm", Some(self.layer_idx), &normalized)?;
-            }
-            normalized
-        } else {
-            k
-        };
-
-        // GQA diagnostic: log Q/K/V dimensions and norms (once per run)
-        if debug_gqa_enabled() {
-            static GQA_LOGGED: std::sync::Once = std::sync::Once::new();
-            GQA_LOGGED.call_once(|| {
-                let q_dims = q.dims();
-                let k_dims = k.dims();
-                let v_dims = v.dims();
-                if let (Ok(q_mean), Ok(k_mean), Ok(v_mean)) = (
-                    q.mean_all().and_then(|m| m.to_scalar::<f32>()),
-                    k.mean_all().and_then(|m| m.to_scalar::<f32>()),
-                    v.mean_all().and_then(|m| m.to_scalar::<f32>()),
-                ) {
-                    tracing::info!(
-                        "GQA shapes - Q: {:?} (mean {:.3}), K: {:?} (mean {:.3}), V: {:?} (mean {:.3})",
-                        q_dims, q_mean, k_dims, k_mean, v_dims, v_mean
-                    );
-                    tracing::info!(
-                        "GQA config - n_heads={}, n_kv_heads={}, head_dim={}, group_size={}",
-                        self.n_heads, self.n_kv_heads, self.head_dim, self.group_size
-                    );
-                }
-            });
-        }
-
-        // Apply rotary embeddings if available (need to handle different K/V head counts)
-        let (q, k) = if let Some(rope) = &self.rope {
-            let position = kv_cache.as_ref().map(|c| c.seq_len).unwrap_or(0);
-
-            // Log ROPE application details (once)
-            if debug_rope_enabled() {
-                static ROPE_LOGGED: std::sync::Once = std::sync::Once::new();
-                ROPE_LOGGED.call_once(|| {
-                    tracing::info!(
-                        "ROPE applied: position={}, q_shape={:?}, k_shape={:?}, head_dim={}",
-                        position,
-                        q.dims(),
-                        k.dims(),
-                        self.head_dim
-                    );
-                });
-            }
-
-            let q_rot = rope.apply(&q, position)?;
-            let k_rot = rope.apply(&k, position)?;
-            if qwen_trace_layer_enabled(self.layer_idx) {
-                qwen_trace_event(
-                    "attention.rope_metadata",
-                    &format!(
-                        "\"layer\":{},\"position\":{},\"head_dim\":{},\"n_heads\":{},\"n_kv_heads\":{}",
-                        self.layer_idx, position, self.head_dim, self.n_heads, self.n_kv_heads
-                    ),
-                );
-                qwen_trace_tensor("attention.q_rope", Some(self.layer_idx), &q_rot)?;
-                qwen_trace_tensor("attention.k_rope", Some(self.layer_idx), &k_rot)?;
-            }
-            (q_rot, k_rot)
-        } else {
-            (q, k)
-        };
-
-        // Update KV cache if provided (store HKV heads, not Hq)
-        // **Performance note**: Borrow references instead of cloning after append.
-        // Candle operations accept both owned and borrowed tensors.
-        let (k_ctx, v_ctx) = if let Some(cache) = kv_cache {
-            cache.append(&k, &v)?;
-            // Borrow from cache - avoids cloning full KV history
-            (&cache.k, &cache.v)
-        } else {
-            // No cache: use freshly computed K/V from this step
-            (&k, &v)
-        };
-
-        // GQA core: expand K/V to Hq heads (repeat along head axis)
-        // We want K,V of shape [B,Hq,Tk,D]. Repeat every KV head group_size times.
-        let t_k = k_ctx.dims()[2];
-
-        // Expand K: [B, HKV, Tk, D] -> [B, Hq, Tk, D]
-        let k_expanded = k_ctx
-            .unsqueeze(2)?                               // [B, HKV, 1, Tk, D]
-            .repeat(&[1, 1, self.group_size, 1, 1])?    // [B, HKV, group, Tk, D]
-            .reshape(&[batch_size, self.n_heads, t_k, self.head_dim])?; // [B, Hq, Tk, D]
-
-        // Expand V: [B, HKV, Tk, D] -> [B, Hq, Tk, D]
-        let v_expanded = v_ctx
-            .unsqueeze(2)?                               // [B, HKV, 1, Tk, D]
-            .repeat(&[1, 1, self.group_size, 1, 1])?    // [B, HKV, group, Tk, D]
-            .reshape(&[batch_size, self.n_heads, t_k, self.head_dim])?; // [B, Hq, Tk, D]
-
-        // Scaled dot-product attention with explicit fp32 handling
-        // For head_dim=128, scale = 1/sqrt(128) ≈ 0.0883883
-        let scale_factor = (self.head_dim as f32).sqrt().recip();
-
-        // Log scale computation once
-        if debug_attn_scale_enabled() {
-            static SCALE_LOGGED: std::sync::Once = std::sync::Once::new();
-            SCALE_LOGGED.call_once(|| {
-                tracing::info!(
-                    "Attention scale: head_dim={}, scale_factor=1/sqrt({})={:.7}",
-                    self.head_dim,
-                    self.head_dim,
-                    scale_factor
-                );
-            });
-        }
-
-        let scores = q.matmul(&k_expanded.transpose(2, 3)?)?;
-
-        // Convert to fp32 for numerically stable computation
-        let scores_f32 = scores.to_dtype(DType::F32)?;
-
-        // Scale in fp32
-        let scores_f32 = scores_f32.affine(scale_factor as f64, 0.0)?;
-
-        // Debug scores before mask
-        dbg_stats("scores pre-mask", &scores_f32)?;
-        dbg_finite("scores pre-mask", &scores_f32)?;
-
-        // Apply causal mask so queries cannot attend to future positions.
-        // When using a KV cache, k includes past tokens, so the mask must
-        // account for the total key length. Single-token decode has no future
-        // key positions, so the mask is all zeros and can be skipped.
-        let total_len = k_expanded.dims()[2];
-        let scores_f32 = if seq_len == 1 {
-            scores_f32
-        } else {
-            // PATCH 5: create_causal_mask now returns [1, 1, Tq, Tk] directly - no need for unsqueeze
-            let mask = self.create_causal_mask(seq_len, total_len, scores_f32.device())?;
-            scores_f32.broadcast_add(&mask)?
-        };
-        if qwen_trace_layer_enabled(self.layer_idx) {
-            qwen_trace_tensor("attention.scores_post_mask", Some(self.layer_idx), &scores_f32)?;
-        }
-
-        // Debug scores after mask and before softmax (critical diagnostics)
-        dbg_stats("scores post-mask", &scores_f32)?;
-        dbg_finite("scores post-mask", &scores_f32)?;
-
-        // Log scores range after mask for layer 0 (user's diagnostic request)
-        if debug_attn_scale_enabled() {
-            static LAYER_LOGGED: std::sync::Once = std::sync::Once::new();
-            LAYER_LOGGED.call_once(|| {
-                if let Ok(flat) = scores_f32.flatten_all()
-                    && let Ok(vals) = flat.to_vec1::<f32>()
-                    && let (Some(&min_val), Some(&max_val)) = (
-                        vals.iter()
-                            .filter(|v| v.is_finite())
-                            .min_by(|a, b| a.partial_cmp(b).unwrap()),
-                        vals.iter()
-                            .filter(|v| v.is_finite())
-                            .max_by(|a, b| a.partial_cmp(b).unwrap()),
-                    )
-                {
-                    tracing::info!(
-                        "Layer 0 scores post-mask range: min={:.6}, max={:.6}",
-                        min_val,
-                        max_val
-                    );
-                }
-            });
-        }
-
-        // PATCH 4: Softmax path verification
-        // Apply max-subtraction for numerical stability before softmax
-        // Compute row-wise max and subtract for stability (explicit max-subtraction)
-        // VERIFIED: axis=3 is correct for [B, H, Tq, Tk] layout - normalizes across keys (Tk)
-        let row_max = scores_f32.max_keepdim(3)?;
-        let scores_stabilized = scores_f32.broadcast_sub(&row_max)?;
-
-        // Log that max-subtraction ran (user's diagnostic request)
-        if debug_attn_scale_enabled() {
-            static MAX_SUB_LOGGED: std::sync::Once = std::sync::Once::new();
-            MAX_SUB_LOGGED.call_once(|| {
-                tracing::info!("Attention: max-subtraction applied for numerical stability");
-            });
-        }
-
-        // Apply softmax (exp then normalize)
-        // VERIFIED: axis=3 is correct - softmax over keys (Tk dimension) in [B, H, Tq, Tk]
-        let attn_weights = candle_nn::ops::softmax(&scores_stabilized, 3)?;
-        if qwen_trace_layer_enabled(self.layer_idx) {
-            qwen_trace_tensor("attention.weights", Some(self.layer_idx), &attn_weights)?;
-        }
-
-        // Tracepoint 4: Attention scores post-softmax (layer-specific)
-        #[cfg(feature = "trace")]
-        {
-            let trace_name = format!("t0/blk{}/attn_scores_softmax", self.layer_idx);
-            bitnet_trace::dump_trace(
-                &trace_name,
-                &attn_weights,
-                Some(0),
-                Some(self.layer_idx as isize),
-                Some("attn_scores_softmax"),
-            )
-            .map_err(BitNetError::from)?;
-        }
-
-        // Debug attention weights and row sums
-        dbg_stats("attn softmax", &attn_weights)?;
-        if debug_attn_enabled() {
-            let sums = attn_weights.sum(3)?;
-            let sums_host: Vec<f32> = sums.flatten_all()?.to_vec1()?;
-            let take = sums_host.iter().take(4).cloned().collect::<Vec<_>>();
-            eprintln!("[dbg] attn row-sums (first 4): {:?}", take);
-        }
-
-        let attn_output = attn_weights.matmul(&v_expanded)?;
-        if qwen_trace_layer_enabled(self.layer_idx) {
-            qwen_trace_tensor("attention.output_heads", Some(self.layer_idx), &attn_output)?;
-        }
-
-        // Reshape and project output
-        let attn_output = attn_output.transpose(1, 2)?.reshape(&[
-            batch_size,
-            seq_len,
-            self.n_heads * self.head_dim,
-        ])?;
-        let attn_output = if let Some(sub_layernorm) = &self.sub_layernorm {
-            let normalized = sub_layernorm.forward(&attn_output)?;
-            if qwen_trace_layer_enabled(self.layer_idx) {
-                qwen_trace_tensor("attention.sub_layernorm", Some(self.layer_idx), &normalized)?;
-            }
-            normalized
-        } else {
-            attn_output
-        };
-
-        let projected = self.apply_linear(&attn_output, &self.o_proj, "o_proj", raw_tensors)?;
-        if qwen_trace_layer_enabled(self.layer_idx) {
-            qwen_trace_tensor("attention.o_proj", Some(self.layer_idx), &projected)?;
-        }
-        Ok(projected)
-    }
+    // `forward` lives in the `attention_forward` module so the projection,
+    // RoPE/cache, GQA, score, softmax, and output-projection responsibilities
+    // stay isolated from QK256 linear dispatch helpers below.
 
     /// Apply linear transformation with QK256 dispatch
     fn apply_linear(
